+++ app/Http/Controllers/DriverController.php
@@ -17,7 +17,7 @@ class DriverController extends Controller
      */
     public function index()
     {
-        return response( DriverResource::collection( Driver::all() ), 200 );
+        return $this->response( 200, DriverResource::collection( Driver::all() ) );
     }
 
     /**
@@ -28,33 +28,7 @@ class DriverController extends Controller
      */
     public function store(Request $request)
     {
-        $request->merge([
-            'slug' => Str::slug($request->input('slug') ?: '')
-        ]);
-
-        list('slug' => $slug, 'name' => $name) = $request->validate( [
-            'slug' => 'nullable|string|max:255|unique:drivers,slug',
-            'name'    => 'required|string|max:255'
-        ] );
-
-        if (empty($slug)) {
-            if (Driver::where('slug', $slug)->exists()) {
-                return response('auto slug exists', 422);
-            }
-        }
-
-        try {
-            DB::beginTransaction();
-            $driver = new Driver();
-            $driver->slug = strtolower($slug);
-            $driver->name = $name;
-            $driver->save();
-            DB::commit();
-            return response('', 201);
-        } catch (\Exception $e) {
-            DB::rollBack();
-            return response('error', 422);
-        }
+        return $this->saveDriver( $request );
     }
 
     /**
@@ -65,7 +39,11 @@ class DriverController extends Controller
      */
     public function show(Driver $driver)
     {
-        return response( DriverResource::collection( $driver ), 200 );
+        if ( ! $driver ) {
+            return $this->response( 404, [], 'Driver Not Found' );
+        } else {
+            return $this->response( 200, new DriverResource( $driver ), 'Found' );
+        }
     }
 
     /**
@@ -77,31 +55,10 @@ class DriverController extends Controller
      */
     public function update(Request $request, Driver $driver)
     {
-        $request->merge([
-            'slug' => Str::slug($request->input('slug') ?: '')
-        ]);
-
-        list('slug' => $slug, 'name' => $name) = $request->validate( [
-            'slug' => 'nullable|string|max:255|unique:drivers,slug',
-            'name'    => 'required|string|max:255'
-        ] );
-
-        if (empty($slug)) {
-            if (Driver::where('slug', $slug)->exists()) {
-                return response('auto slug exists (' . $slug . ')', 422);
-            }
-        }
-
-        try {
-            DB::beginTransaction();
-            $driver->slug = strtolower($slug);
-            $driver->name = $name;
-            $driver->save();
-            DB::commit();
-            return response('', 204);
-        } catch (\Exception $e) {
-            DB::rollBack();
-            return response('error', 422);
+        if ( ! $driver ) {
+            return $this->response( 404, [], 'Driver Not Found' );
+        } else {
+            return $this->saveDriver( $request, false, $driver );
         }
     }
 
@@ -113,7 +70,50 @@ class DriverController extends Controller
      */
     public function destroy(Driver $driver)
     {
-        $driver->delete();
-        return response('', 204);
+        if ( ! $driver ) {
+            return $this->response( 404, [], 'Driver Not Found' );
+        } else {
+            try {
+                $driver->delete();
+
+                return $this->response( 204, [], 'Successfully deleted' );
+            } catch ( \Exception $e ) {
+                return $this->response( 500, [ 'error' => $e->getMessage() ], 'Error while processing' );
+            }
+        }
+    }
+
+    private function saveDriver( Request $request, bool $isCreated = true, Driver $driver = null ) {
+        $request->merge([
+            'slug' => Str::slug($request->input('slug') ?: '')
+        ]);
+
+        list('slug' => $slug, 'name' => $name) = $request->validate( [
+            'slug' => 'nullable|string|max:255|unique:drivers,slug',
+            'name'    => 'required|string|max:255'
+        ] );
+
+
+        if (empty($slug)) {
+            if (Driver::where('slug', $slug)->exists()) {
+                return response('auto slug exists (' . $slug . ')', 422);
+            }
+        }
+
+        try {
+            DB::beginTransaction();
+            if ( $isCreated ) {
+                $driver = new Driver();
+            }
+            $driver->slug = $slug;
+            $driver->name = $name;
+            $driver->save();
+            DB::commit();
+
+            return $this->response( 201, [], 'Successfully processed' );
+        } catch (\Exception $e) {
+            DB::rollBack();
+            return $this->response( 500, [ 'error' => $e->getMessage() ], 'Error while processing' );
+        }
     }
 }
