+++ app/Http/Controllers/FuelEntryController.php
@@ -6,7 +6,6 @@ use App\Http\Resources\FuelEntryResource;
 use App\Models\FuelEntry;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use mysql_xdevapi\Exception;
 
 class FuelEntryController extends Controller {
     /**
@@ -15,7 +14,7 @@ class FuelEntryController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
-        return response( FuelEntryResource::collection( FuelEntry::all() ), 200 );
+        return $this->response( 200, FuelEntryResource::collection( FuelEntry::all() ) );
     }
 
     /**
@@ -26,23 +25,7 @@ class FuelEntryController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store( Request $request ) {
-        list('fuel_type' => $fuel_type, 'amount' => $amount) = $request->validate( [
-            'fuel_type' => 'required|in:gasoline,gas_oil',
-            'amount'    => 'required|int'
-        ] );
-
-        try {
-            DB::beginTransaction();
-            $fuelEntry = new FuelEntry();
-            $fuelEntry->fuel_type = $fuel_type;
-            $fuelEntry->amount = $amount;
-            $fuelEntry->save();
-            DB::commit();
-            return response('', 201);
-        } catch (\Exception $e) {
-            DB::rollBack();
-            return response('error', 422);
-        }
+        return $this->saveFuelEntry( $request );
     }
 
     /**
@@ -53,7 +36,11 @@ class FuelEntryController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function show( FuelEntry $fuelEntry ) {
-        return response( FuelEntryResource::collection( $fuelEntry ), 200 );
+        if ( ! $fuelEntry ) {
+            return $this->response( 404, [], 'Fuel Entry Not Found' );
+        } else {
+            return $this->response( 200, new FuelEntryResource( $fuelEntry ), 'Found' );
+        }
     }
 
     /**
@@ -65,21 +52,10 @@ class FuelEntryController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update( Request $request, FuelEntry $fuelEntry ) {
-        list('fuel_type' => $fuel_type, 'amount' => $amount) = $request->validate( [
-            'fuel_type' => 'required|in:gasoline,gas_oil',
-            'amount'    => 'required|int'
-        ] );
-
-        try {
-            DB::beginTransaction();
-            $fuelEntry->fuel_type = $fuel_type;
-            $fuelEntry->amount = $amount;
-            $fuelEntry->save();
-            DB::commit();
-            return response('', 204);
-        } catch (\Exception $e) {
-            DB::rollBack();
-            return response('error', 422);
+        if ( ! $fuelEntry ) {
+            return $this->response( 404, [], 'Fuel Entry Not Found' );
+        } else {
+            return $this->saveFuelEntry( $request, false, $fuelEntry );
         }
     }
 
@@ -91,7 +67,40 @@ class FuelEntryController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy( FuelEntry $fuelEntry ) {
-        $fuelEntry->delete();
-        return response('', 204);
+        if ( ! $fuelEntry ) {
+            return $this->response( 404, [], 'Fuel Entry Not Found' );
+        } else {
+            try {
+                $fuelEntry->delete();
+
+                return $this->response( 204, [], 'Successfully deleted' );
+            } catch ( \Exception $e ) {
+                return $this->response( 500, [ 'error' => $e->getMessage() ], 'Error while processing' );
+            }
+        }
+    }
+
+    private function saveFuelEntry( Request $request, bool $isCreated = true, FuelEntry $fuelEntry = null ) {
+        list( 'fuel_type' => $fuelType, 'amount' => $amount ) = $request->validate( [
+            'fuel_type' => 'required|in:gasoline,gas_oil',
+            'amount'    => 'required|int'
+        ] );
+
+        try {
+            DB::beginTransaction();
+            if ( $isCreated ) {
+                $fuelEntry = new FuelEntry();
+            }
+            $fuelEntry->fuel_type = $fuelType;
+            $fuelEntry->amount    = $amount;
+            $fuelEntry->save();
+            DB::commit();
+
+            return $this->response( 201, [], 'Successfully processed' );
+        } catch ( \Exception $e ) {
+            DB::rollBack();
+
+            return $this->response( 500, [ 'error' => $e->getMessage() ], 'Error while processing' );
+        }
     }
 }
