+++ app/Http/Controllers/FuellingController.php
@@ -7,96 +7,107 @@ use App\Models\Fuelling;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
-class FuellingController extends Controller
-{
+class FuellingController extends Controller {
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        return response( FuellingResource::collection( Fuelling::all() ), 200 );
+    public function index() {
+        return $this->response( 200, FuellingResource::collection( Fuelling::all() ) );
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
+     *
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
-        return $this->save_fuelling($request);
+    public function store( Request $request ) {
+        return $this->saveFuelling( $request );
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Fuelling  $fuelling
+     * @param \App\Models\Fuelling $fuelling
+     *
      * @return \Illuminate\Http\Response
      */
-    public function show(Fuelling $fuelling)
-    {
-        return response( FuellingResource::collection( $fuelling ), 200 );
+    public function show( Fuelling $fuelling ) {
+        if ( ! $fuelling ) {
+            return $this->response( 404, [], 'Fuelling Not Found' );
+        } else {
+            return $this->response( 200, new FuellingResource( $fuelling ), 'Found' );
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Fuelling  $fuelling
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\Fuelling $fuelling
+     *
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Fuelling $fuelling)
-    {
-        return $this->save_fuelling($request, false, $fuelling);
+    public function update( Request $request, Fuelling $fuelling ) {
+        return $this->saveFuelling( $request, false, $fuelling );
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Fuelling  $fuelling
+     * @param \App\Models\Fuelling $fuelling
+     *
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Fuelling $fuelling)
-    {
-        $fuelling->delete();
-        return response('', 204);
+    public function destroy( Fuelling $fuelling ) {
+        if ( ! $fuelling ) {
+            return $this->response( 404, [], 'Fuelling Not Found' );
+        } else {
+            try {
+                $fuelling->delete();
+
+                return $this->response( 204, [], 'Successfully deleted' );
+            } catch ( \Exception $e ) {
+                return $this->response( 500, [ 'error' => $e->getMessage() ], 'Error while processing' );
+            }
+        }
+
     }
 
-    private function save_fuelling (Request $request, bool $isCreated = true, Fuelling $fuelling = null) {
+    private function saveFuelling( Request $request, bool $isCreated = true, Fuelling $fuelling = null ) {
         $request->validate( [
             'driver_id' => 'required|int|exists:App\Models\Driver,id',
             'car_id'    => 'required|int|exists:App\Models\Car,id',
             'fuel_type' => 'required|in:gasoline,gas_oil',
-            'amount' => 'required|int',
+            'amount'    => 'required|int',
         ] );
-
 
         try {
 
             DB::beginTransaction();
 
-            if ($isCreated) {
+            if ( $isCreated ) {
                 $fuelling = new Fuelling();
             }
-            $fuelling->driver_id = $request->input('driver_id');
-            $fuelling->car_id = $request->input('car_id');
-            $fuelling->fuel_type = $request->input('fuel_type');
-            $fuelling->amount = $request->input('amount');
+            $fuelling->driver_id = $request->input( 'driver_id' );
+            $fuelling->car_id    = $request->input( 'car_id' );
+            $fuelling->fuel_type = $request->input( 'fuel_type' );
+            $fuelling->amount    = $request->input( 'amount' );
 
             $fuelling->save();
 
             DB::commit();
 
-            return response(FuellingResource::collection( Fuelling::all() ), 201);
+            return $this->response( 201, [], 'Successfully processed' );
 
-        }  catch (\Exception $e) {
+        } catch ( \Exception $e ) {
 
             DB::rollBack();
 
-            return $this->responseError('Sorry, but something went wrong. Try again.', 500);
+            return $this->response( 500, [ 'error' => $e->getMessage() ], 'Error while processing' );
 
         }
     }
